Refactored audio classes into an audio model module
